requirements.txt
tiktoken==0.7.0
openai==1.34.0
google-cloud-aiplatform==1.55.0
google-cloud-documentai==2.29.0
pillow==10.3.0
PyYAML==6.0.1
python-dotenv==1.0.1
tkintertable==1.3.3